feat: 1st pass wiring up Bolt custom Grid and Cell tags using new Class based extension system; adding workaround to autoloading, removing old extension tags being replaced
