+++ packages/core-php/composer.json
@@ -22,7 +22,8 @@
   "autoload": {
     "psr-0": {
       "Drupal\\Core\\": "src/libs/",
-      "Drupal\\Core\\Template\\": "src/libs/"
+      "Drupal\\Core\\Template\\": "src/libs/",
+      "Bolt\\Layout\\": "src/libs/"
     },
     "psr-4": {
       "Bolt\\": "src/"
